backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
graphene==3.3
starlette-graphene3==0.6.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.13.0
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1